challenge2/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf2>=3.0.0
python-docx>=1.0.0
pydantic>=2.0.0
numpy>=1.21.0
requests>=2.25.0
beautifulsoup4>=4.10.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
